perspective-sql/src/main/java/org/meridor/perspective/sql/impl/function/SignFunction.java
package org.meridor.perspective.sql.impl.function;

import org.springframework.stereotype.Component;

import java.util.List;
import java.util.Set;

import static org.meridor.perspective.sql.impl.function.FunctionUtils.*;

@Component
public class SignFunction implements Function<Double> {

    @Override
    public Set<String> validateInput(List<Object> args) {
        return oneOf(
                args,
                argsCount(1),
                isNumber(0)
        );
    }

    @Override
    public Class<Double> getReturnType() {
        return Double.class;
    }

    @Override
    public FunctionName getName() {
        return FunctionName.SIGN;
    }

    @Override
    public Double apply(List<Object> objects) {
        return Math.signum(Double.valueOf(String.valueOf(objects.get(0))));
    }

    @Override
    public String getSignature() {
        return "SIGN(X)";
    }

    @Override
    public String getDescription() {
        return "Returns the sign of the argument as -1, 0, or 1, depending on whether X is negative, zero, or positive.";
    }

}
